refactor: sex toggle 개별 파일로 분리
