+++ apps/client/src/components/index.ts
@@ -6,3 +6,4 @@ export { default as Header } from './Header';
 export { default as MockCalculatePage } from './MockCalculatePage';
 export { default as TestCalculatePage } from './TestCalculatePage';
 export { default as FormItem } from './FormItem';
+export { default as SexToggle } from './SexToggle';
